Genasys

Documentation for Genasys.

Genasys.grid2Dkron

Genasys.grid3Dkron

Genasys.laplacian

Genasys.laplacian

Genasys.read

Genasys.stencilmat

Genasys.stencilmat

Genasys.stencilmat

Genasys.write

Genasys.grid2Dkron — Method

grid2Dkron(Ax, Ay)

Returns a sparse matrix generated from input X and Y matrices with a kronecker product on a 2D grid.

Examples

julia>  Ax = sprand(2, 2, 0.5)
julia>  Ay = sprand(3, 3, 0.75)
julia>  grid2Dkron(Ax, Ay)
6×6 SparseMatrixCSC{Float64, Int64} with 14 stored entries:
 0.375082   ⋅        0.137573   ⋅        0.34465    ⋅
  ⋅        0.789502   ⋅        0.137573   ⋅        0.34465
 0.523429   ⋅        1.04008    ⋅         ⋅         ⋅
  ⋅        0.523429   ⋅        1.4545     ⋅         ⋅
  ⋅         ⋅        0.86456    ⋅        0.234737   ⋅
  ⋅         ⋅         ⋅        0.86456    ⋅        0.649157

Genasys.grid3Dkron — Method

grid3Dkron(X, Y, Z)

Returns a sparse matrix generated from input X, Y and Z matrices with a kronecker product on a 3D grid.

Examples

julia>  Ax = sprand(2, 2, 0.5)
julia>  Ay = sprand(3, 3, 0.75)
julia>  Az = sprand(3, 3, 0.4)
julia>  grid3Dkron(Ax, Ay, Az)
18×18 SparseMatrixCSC{Float64, Int64} with 66 stored entries:
⎡⡻⢎⡢⠑⠄⠀⠀⠀⠀⎤
⎢⠈⠪⡛⣤⡀⠀⠀⠀⠀⎥
⎢⠀⠀⠈⠈⠱⣦⠢⡐⢄⎥
⎢⠀⠀⠀⠀⠈⠢⡻⢆⡀⎥
⎣⠀⠀⠀⠀⠀⠀⠈⠊⠛⎦

Genasys.laplacian — Method

laplacian(n₁, n₂, n₃)

Returns a sparse matrix for the Laplacian on a 3D rectangular grid of size n₁ × n₂ × n₃ .

Examples

julia>  laplacian(2, 2, 2)
8×8 SparseArrays.SparseMatrixCSC{Float64, Int64} with 32 stored entries:
  6.0  -1.0  -1.0    ⋅   -1.0    ⋅     ⋅     ⋅
 -1.0   6.0    ⋅   -1.0    ⋅   -1.0    ⋅     ⋅
 -1.0    ⋅    6.0  -1.0    ⋅     ⋅   -1.0    ⋅
   ⋅   -1.0  -1.0   6.0    ⋅     ⋅     ⋅   -1.0
 -1.0    ⋅     ⋅     ⋅    6.0  -1.0  -1.0    ⋅
   ⋅   -1.0    ⋅     ⋅   -1.0   6.0    ⋅   -1.0
   ⋅     ⋅   -1.0    ⋅   -1.0    ⋅    6.0  -1.0
   ⋅     ⋅     ⋅   -1.0    ⋅   -1.0  -1.0   6.0

Genasys.laplacian — Method

laplacian(n₁, n₂)

Returns a sparse matrix for the Laplacian on a 2D rectangular grid of size n₁ x n₂ .

Examples

julia>  laplacian(1, 1)
sparse([1], [1], [4.0])

Genasys.read — Method

read(filename)

Reads a sparse matrix (in a the MatrixMarket format_header) from a file

Examples

julia>  mat = read("mymatrix.mtx")

Genasys.stencilmat — Method

stencilmat(stencil, m, n, l)

Returns a sparse matrix generated from a stencil on a 3D grid with m × n × l points.

Examples

julia>  stencilmat([-1 2 -1], 2, 2, 2)
8×8 SparseArrays.SparseMatrixCSC{Float64, Int64} with 32 stored entries:
  6.0  -1.0  -1.0    ⋅   -1.0    ⋅     ⋅     ⋅
 -1.0   6.0    ⋅   -1.0    ⋅   -1.0    ⋅     ⋅
 -1.0    ⋅    6.0  -1.0    ⋅     ⋅   -1.0    ⋅
   ⋅   -1.0  -1.0   6.0    ⋅     ⋅     ⋅   -1.0
 -1.0    ⋅     ⋅     ⋅    6.0  -1.0  -1.0    ⋅
   ⋅   -1.0    ⋅     ⋅   -1.0   6.0    ⋅   -1.0
   ⋅     ⋅   -1.0    ⋅   -1.0    ⋅    6.0  -1.0
   ⋅     ⋅     ⋅   -1.0    ⋅   -1.0  -1.0   6.0

Genasys.stencilmat — Method

stencilmat(stencil, m, n)

Returns a sparse matrix generated from a stencil on a 2D grid with m × n points.

Examples

julia>  stencilmat([-1 2 -1], 3, 2)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 20 stored entries:
  4.0  -1.0    ⋅   -1.0    ⋅     ⋅
 -1.0   4.0  -1.0    ⋅   -1.0    ⋅
   ⋅   -1.0   4.0    ⋅     ⋅   -1.0
 -1.0    ⋅     ⋅    4.0  -1.0    ⋅
   ⋅   -1.0    ⋅   -1.0   4.0  -1.0
   ⋅     ⋅   -1.0    ⋅   -1.0   4.0

Genasys.stencilmat — Method

stencilmat(stencil, n)

Returns a sparse matrix generated from a stencil on a 1D grid with n points.

Examples

julia>  stencilmat([-1 2 -1], 4)
4×4 SparseArrays.SparseMatrixCSC{Float64, Int64} with 10 stored entries:
  2.0  -1.0    ⋅     ⋅
 -1.0   2.0  -1.0    ⋅
   ⋅   -1.0   2.0  -1.0
   ⋅     ⋅   -1.0   2.0

Genasys.write — Method

write(filename, matrix)

Writes a sparse matrix to a file in the matrix market format

Examples

mat = sprand(n, n)
julia>  write("mymatrix.mtx", mat)